+++ internal/server/transport/wsmux.go
@@ -270,10 +270,10 @@ func (s *WsMuxTransport) channelHandler(g *wsMuxGen) {
 	for {
 		select {
 		case <-g.ctx.Done():
-			_ = s.controlChannel.Get().WriteMessage(websocket.BinaryMessage, []byte{utils.SG_Closed})
+			_ = writeControl(s.controlChannel.Get(), []byte{utils.SG_Closed})
 			return
 		case <-g.reqNewConnChan:
-			err := s.controlChannel.Get().WriteMessage(websocket.BinaryMessage, []byte{utils.SG_Chan})
+			err := writeControl(s.controlChannel.Get(), []byte{utils.SG_Chan})
 			if err != nil {
 				s.logger.Error("failed to send request new connection signal. ", err)
 				go s.Restart()
@@ -281,7 +281,7 @@ func (s *WsMuxTransport) channelHandler(g *wsMuxGen) {
 			}
 
 		case <-ticker.C:
-			err := s.controlChannel.Get().WriteMessage(websocket.BinaryMessage, []byte{utils.SG_HB})
+			err := writeControl(s.controlChannel.Get(), []byte{utils.SG_HB})
 			if err != nil {
 				s.logger.Errorf("failed to send heartbeat signal. Error: %v.", err)
 				go s.Restart()
@@ -394,7 +394,7 @@ func (s *WsMuxTransport) tunnelListener(g *wsMuxGen) {
 				select {
 				case g.tunnelChannel <- session: // ok
 				default:
-					s.logger.Warnf("forwarded port: the queue is full, dropping a client from %s", conn.LocalAddr().String())
+					s.logger.Warnf("forwarded port: the queue is full, dropping a client from %s", conn.RemoteAddr().String())
 					conn.Close()
 				}
 			}
@@ -651,7 +651,7 @@ func (s *WsMuxTransport) acceptLocalConn(g *wsMuxGen, listener net.Listener, rem
 				}
 
 			default: // channel is full, discard the connection
-				s.logger.Warnf("forwarded port: the queue is full, dropping a client from %s", tcpConn.LocalAddr().String())
+				s.logger.Warnf("forwarded port: the queue is full, dropping a client from %s", tcpConn.RemoteAddr().String())
 				s.limits.release()
 				conn.Close()
 			}